Cleanup in FileChannelImpl
2006-01-04  Andreas Tobler  <a.tobler@schweiz.ch>

        * libraries/clib/nio/gnu_java_nio_channels_FileChannelImpl.c
        (Java_gnu_java_nio_channels_FileChannelImpl_implTruncate): Use the same
        type for bytes_written as in other places.

+++ libraries/clib/nio/gnu_java_nio_channels_FileChannelImpl.c
@@ -384,7 +384,7 @@ Java_gnu_java_nio_channels_FileChannelImpl_implTruncate (JNIEnv * env,
 {
   int native_fd;
   jlong file_size;
-  int bytes_written;
+  ssize_t bytes_written;
   jlong save_offset, new_offset;
   char data;
   int result;
